requirements.txt
appdirs==1.4.4
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.10.0
black==20.8b1
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
dj-database-url==0.5.0
Django==4.1.5
django-cors-headers==3.7.0
django-heroku==0.3.1
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.25.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.17.3
lxml==4.6.3
Markdown==3.4.1
markdown2==2.4.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.8.1
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.2.0
pyparsing==3.0.6
pyquery==1.4.3
pyrsistent==0.19.3
pytz==2021.1
PyYAML==6.0
readtime==1.1.1
regex==2021.4.4
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
